Adding custom database export/dump script.
In the continuing effort to build more immediate start-up (OOB) functionality into the project, we should commit some test database data.
What better data to use than a snapshot of production!
We should only need a handful from several categories in the dataset, but:
 - it is a relatively small dataset currently,
 - and it would take a great amount of time to determine and prune it all down to a minimum viable dataset,

So I made this script that essentially runs the `mysqldump` command a couple times and does some file massaging.
This ware carefully done to avoid certain tables with possibly sensitive user data, all the while making this a convenient utility in case we can use it to update the dataset in this repository, or anywhere else!

We can include the database data inside of this repository for the first time once I make sure we won't have a quick update available soon, which is also why I am committing this script update first in caseit can be used soon enough.