Adding search trigger functionality to the search field.
Pushing this early for the team to preview the design and current functionality.

Tried to keep some nice OO/encapsulation concepts I saw elsewhere in the code for low coupling, better composibility, easier maintenance through more concise code, and decent configuratbility when adding these widgets!

Adding a base util function for those frequent widget config setting reads.

If the search field widgets is needed elsewhere, we can pull apart and generalize some of the code into more configuration points.
The CSS only currently works for the one instance, and is not generalized well enough yet.

Also made the progress bar widget that felt separate enough and helped ease development by separating concerns early.
Made sure not to include time concepts in it, and instead in the search field waiting pipeline.

Updated javascript include file with specific ordering.
This is going to be an annoying part of the process..