Extending the escape key-clear functionality to the search mechanism as well. It is also nice and dynamic, detecting if there is anything to clear in the search field when pressed. When not, the next 'phase' can be controlled by config, either directly toggling the entire drawer again, or with that as a 3rd step, the 2nd would be to only clear currently searched results next.
This required making another change in the drawer code that I missed when starting to implement the search results feature.
I added another state for the lower pane to be in, which helps the escape key-clear code work.
Also went ahead and simplified the set content method to work with strings of, or actual, DOM nodes, like I had introduced.  Allows me to be a bit more efficient in the handlers that update the drawer pane.
Fixed some contentType variables without the usual prefix.  Hopefully that makes things better.

Along with this, I added a clear 'x' icon to the search field.
I made a second class that I thought worked better, since I have re-figured out CSS positions to work for me again with what I did in a future search-related commit, and since I have been suggested to avoid using floats.
Made it sticky for better re-use in the scrolling search results area too.  Changed the class to use mine now.
Found the right event for the clear button to update visibility on, so that is nice and spiffy (provides the best/immediate UI experience/communication)!

A bit more refactoring the search as I add complexity to the search field.
Resulted in me allowing the enter key press to also execute search when in incremental mode, since this may be useful for certain timing configurations, and at all for another decent UI experience.
Made sure it clears the incremental state well too!