Making a change to try to focus the search field by default, for easier user experience. Should help in the cases where it would be the next intended action, and not hurt others, like when the user just wants to click a category icon. This would be wanted every time the search field becomes visible, like when the drawer dialog box is expanded.
Had to rearrange some drawer code for it to work, like placing it after any expand calls.
In the earlier larger function, not sure why I had to move that expand block near the end of the function.  I think it just makes sense to do it after any potential DOM preparation code, and serves as a nice convention to have it at the near-end.