Alright, adding the core searching functionality.
Added a searcher object that controls the Fuse searching library.
This will handle its settings and any events we need it with.  Glue code!
It uses some search data parameters that I barely chose.  These will probably have to be tweaked in the future!!

In order to use the library, I had to add a zMap hook, since the markers were not available during the initial flow of object construction / page load process.
This will trigger the instantiation of the search library.
Was able to keep the zMap dependency high for low coupling!

Fixed the search field's event to simply attach the current query entered as the only parameter for now.

Added a search header I thought would be useful displaying the current query that is searched, and how many results are returned.
Made sure the marker list view and entry code were separated enough when adding their subclasses.