Handler that logically links related components, now has a setting for which marker list view style you want.
Added a new list view and entry objects that show more search stats.
This is useful as a developer, and even as a user so they can see the magnitude of how relevant search results are absolutely, and relatively to reach other.
Configurable!
This involved changing / breaking apart some super class code that can be overridden in pieces now.
Tried simplifying some methods.
A bit weird with the marker parameter now and calling the inherited/overridden settings methods.  Maybe this is why `opts` more loose parameters are also recommended, for better composability?
For the constructors, started simply dealing with the only marker parameter for now as well.
Also found the funny nature of our marker data objects, about what properties they have / are named, and when they have them / are loaded, so I try to support both.  They could be cleaned up in the future to only have a single interface!

These score indicators are also configurable, almost eliminating the need for separate classes, but using the plainer version could save on performance, so I kept it.  Using the search stats version also expects a search entry to be passed in, and not only marker data object, so that was also a reason for having separate interfaces, although not a large one.
I added both a visual and textual indicator that are each useful, and tried really hard to find styles that would work given our dialog constraints and existing elements.
Added doc for the progress bar, since that is now also involved in one indicator.
I changed the text score to be inverted, then a percentage, and finally rounded to 1 single decimal place, as a nice balance of accuracy, human understanding, and presentation.
Rounding isn't great in pure JS, so I added someone's general helper function to help with this.  It works well.

Also added highlighting!
Wasn't the most trivial feature, but came up with a view algorithm that works.  It's similar to the one I found in the GitHub issue thread.
Made it a general function that could be used in other places, instead of just in the marker search entries.
Grouping mechanism for the different property's indices may not completely be needed for the API, but it at least puts it into a slightly more easily consumable object format for me to parse.

Cleaned some interfaces as I'm going.

Throwing all of the styles I worked so hard on, for the old and new components.
There's a calc bit for the marker entry width that I wasn't happy with.  Could probably be improved on later.

P.S. Yes, my search parameter has a weird name.  Big whup!  Wanna fight about it? :)