Adding basic functionality clicking marker search results to navigate to that spot on the map.
This does not feel complete, and may even have some slight inconsistency bugs, but I want to push out these changes for review so I can get some feedback on it, since there has not been much of any detailed design requirements specified for this feature.

I did make nifty highlight, cursor change, and click visual effects for the marker search list entries!

Continuing to keep the zMap dependency nested deeply out of child objects!
Fixed up all of the related objects to help pass this through.  Also tried to keep the interfaces reasonable for each one, using different names as the function is passed down that relate only to the next object, in case a portion of the objects are used elsewhere!
Enhanced the open marker function to use the cache.  #nomoremarkeriteration