Changed to handling passwords in a more 'pure' way.
Became a problem when I was trying to reset passwords, which would need to have a consistent algorithm.

I felt it as simpler using the password as-is, instead of escaping it, since we do and should not present it to the user or anyone else, or otherwise store it in its raw form.

Since we don't want to try to perform a password migration, since this isn't a critical issue, and especially since they are only known 'one-way' (hashed), there is no easy method to do this, so I change the login method to accept both forms for now.

Naïvely speaking, this should not reduce the complexity and safety/security of the passwords much at all, since we should all be using a large amount of characters for, and unique passwords for each account we have.