Looking through the search issue with not being able to find the recommended shrine.
I feel it is related to the data, so I found some interesting pieces that were not complete from the previous backups given to me, and now with my own prod db access, was able to get a refreshed, more filled version.

I wrote plenty of queries, expressions, and commands to get most of the data I needed, and sanitize the rest.
For this, I enhanced and refactored the dev db export script.
There are now tons of options to choose from and debug with as well!
Also threw in some nifty code folding since there is plenty there now.
I realized late this essentially turned into a couple of steps in a custom-built build tool.
Should now have result messages, more streamlined output, debug wait, verbose, captured/hooked user input, intelligent cleaning, and colors!!
Kept the old features like dry run testing, and now works pretty well now that I've learned to use MySQL's config editor feature.

As far as the data goes, I removed the assumption that marker_tab is now needed based on the query that gets markers for the map UI.
I also generate a bunch of user accounts with fake data (sanitized!!) so any links that are also needed for that main monster marker query also exist.
These are mainly from the marker table now, which seems to be the largest source of user ids, with any stragglers from marker_tab.
I do not limit visibility so as to incorporate any possible future scenarios a developer might want.
Some others pieces of data were updated too, like the troublesome markers were replaced with better ones, like the missing shrine, and/or it had the visibility toggled off, so now we have a more authoritative version checked in.
However, there are these hack messages sprinkled into the data that maybe should not be there.  Anyway, we can discuss the file and improve upon it more now that it has some source control, instead of just messing around with production data O_o.

Changed the names to also match the production database convention to be helpful.

It's a pretty nice system now that can be run as needed in the future, especially from a local PC with an SSH tunnel and further setup mysql login configuration!

Updated the development and install sections related to database work on how to more completely utilize this script's features, and some other useful commands and shortcuts/tips.