Finished lost/reset password feature by adding the email functionality. Obviously this needs any, and possibly per-node configuration based on deployment details, so I have added configuration fields for sending mail. I left in some sample values to help give an idea of what may be used for each, along with others already filled in with their actual value for the project that is less likely to change. I also cleaned up some of the other fields in the same way. The body of the message could be large, so I moved this to be loaded from a file instead.
Used SwiftMailer based on recommendation over PHP's built-in mail function.  PHPMailer was also a recommendation.  So this needs a dependency install.
Heard we use PHP 7 in production and/or stage.  Not sure how I got started with 5.  Maybe a nice common denominator.  Also thinking that's the version of php/php-fpm that came with my Mac too.

Added more layers to the API between success and fails.  There must be a better way instead of so many nested if statements!! :P  but we'll wait until a back-end API refactor for that..

Having the mail code in a separate file keeps it nice and clean too, plus keeps a lot of related code together in case you need to modify it in the future.

So far we're just using Google's Gmail servers in production, so I'll leave SSL ((START)TLS?) hard-coded enabled on for now, and that seems good to be secure by default too! :)