Setting categories to all be enabled by default.
Was turned off when the category menu was simply re-done.

Since they were redone in a simpler selection style, but we originally had an algorithm for a "smart" style, there were 2 category checked properties to track.
I am additionally setting the second to match the first without removing any of this underlying diverse functionality.

Updated library general function to accept both singular and plural forms now to help with the category button initial state setting from multiple sources.

Since the objects know how to deal with categories already, just moved that initial state logic inside of the button class and cleaned up the menu a bit.

We have 2 APIs and code flow for fetching and category data in different formats.  Probably don't need both, but I left them in and fixed both to work.
Surfaced the checked information from the API for this.

Had a default flag in the category menu set that wasn't being used.  There aren't much cases for this since we have database information dictating when categories should be on by default or not, so I simply too this out.

When working with a category load method, simplified it a bunch so it's easier to look at and work with, more efficient with less lines too.