Adding Search UI (with Login button) to the mobile UI.
Made the mobile layer class extend from the desktop version, so this should help future coding, especially demonstrating how I fixed an old bug again concerning the category type typo.
Gave setContent some surrounding hooks so I could re-use the intersecting core implementation.
Tried to move and keep the code exactly.
Wanted to fix the animate function from being painfully slow, and maybe inefficient, but not now.
Was finally able to have the default focus I wanted by disabling some click handler!  Added a comment just in case.
For some reason thought people might not want it on page load, but it also doesn't work yet; maybe later.

Was able to clean a fair amount of layout code into new widgets, since I needed to re-use them anyway for mobile, and in a slightly different manner!  Not the most cohesive iteration, but still pretty good for now!
Tried putting all this in a nice and proper Leaflet control as might be best, so keeping the abstraction layers sort of clean here too.

Had to add some CSS for the mobile header bar since I wasn't able to quickly re-use anything I found from Leaflet.  'leaflet-bar' only affected the button link's background, but I wanted the entire bar to have it.  It seems this is ample styling for the Zoom control, just not enough for mine. This wasn't a new layer control, more of a toolbar that interacts with the layer control, but separately if you look at the design.  It also wasn't a tooltip either.

So the design excluded this detail.  I had to move the ads under the bar so you can interact with it.  This can be iterated upon.