Kept category selection event listeners, and now fixed to re-allow the marker selection details pane to be displayed, by clearing the main content area even better now.  Just wait until we use more widgets!!
Opening a marker on mobile and hitting escape (not expecting vast majority of mobile users to do this, and may not even be possible if tablets register as desktop to Leaflet) opens the view even larger when you might think it would close.  This is correct, as it resets and shows the default view first, and will close after that state.  If we implement a state stack, then we might want this functionality then, and could be easier to change as well.

Random bit of cleanup.