Adding tool to help us craft migrations easily enough for this (currently :P) non-Rails project. Funny using bundler shows through dependencies that it brings in Rails packages anyway O_o, but I guess that's better than actually running them if you don't need them for the web site.
We're going to see how this works without checking in (ignoring) the schema file for now.
We already have some structure captured for development, and production data backups for that environment, but that file would be useful to keep to track which migrations have been run.  Maybe no need to track in the code repository, and just add it to a backup routine.

Other various doc updates with notes that were recently discussed or performed.