Moving category button to header menu above with logo. Moved the code out from the category menu appropriately.
Deduplicated some logo code in the Bottom Layers class.
Added a useful class for the category button in case it is needed later, along with the consistency with the now matching desktop view adding the header parent element too.

Removed some logo class names that were getting in that way, and didn't seem to add much value.