Bugfix and Small AJAX QoL
For some reason a wrong key was being used to reference a database field in row data, so I fixed it with snake case.  Maybe with me using PHP 8 now it found the error, or something in my local PHP configuration..
Also added a variable to use for changing the content type so PHP errors would appear better.