Added migration for missing changelog entries
that were added in c0b1fb226259310794ac9296a2c5294cf506bf11.
Updated install and dev doc with useful commands and samples,
along with the gemfiles I was updating and getting to work to create these changes,
and including sample SQL data fixes not part of the mentioned earlier commit.