Add PHP MySQLi check
for one situation when I was missing the configure flags when switching to a lower version of PHP, thinking it would be enabled by default..
Forgot to include the new SELinux rules when first running the web server to the new system PHP-CGI/FPM socket path.
Also through page load testing found and removed erroneous check int he get_games API that was added because it matched other code.