Add quick command script
for working with the database server as it relates to the project, with developed features and options taken from the export script, converted to fish shell syntax, and could be used to simplify and clean that, and any other script.
Updated doc as I was refreshing my database.
Made batch import commands not a script as to retain promoting caution..