Add layers/drawer control config
Also felt like a fix for the effective min or normal height of the control to take the max of each possible value, so the minimum is 250 instead of the strange previous behavior that also grew but also could result in essentially hidden and unusable content areas.
Updated Debug/trace/verbose code with config as well.