Updated db migration gem
Actually updated recently, so should have fixed both the missing `schema_file` function error, and the deprecation warning of it being removed in Rails 8, along with the ability to run tasks without an active configuration, which does not apply here.

Removing the related documentation since it is no longer necessary.