Finishing preselect categories configuration feature
Wanted to complete the majority of refactoring changes from previously.
Decided those changes should be at least minor, instead of at the patch level.  Just don't need to advertise it a lot.

This allowed me to have more freedom with testing, and even made an additional fix for the marker popup content not having check marks after interacting with the related button, and even the completed check mark at all.  Among avoiding needing more refactoring there into a class to remove redundancy and improve maintenance efforts that I have always thought of for now, chose the solution that does not require modifying `style.css`, to save effort, and not even sure if I can customize basic symbol names in their interface.  Another patch version for this too.

Added manual link capabilities to aid in using the new feature for users that do not control it with other possible methods.  Decided to add to the bottom after some iterating thinking others may want to prioritize the default, historical presentation of the control, and may even relate to its possibly less used frequency too.  When making the new styling section for this, I left the comments showing relation to the source and my historical iteration progress; probably remove later to clean once we have it stored in the history a bit.  Added a patch version for this.

Fixed problem introduced earlier when more data is returned from the database with categories, some inactive/unused ones need hiding, but this originally applied by default to other users of the category menu for tasks unrelated to category control, which I also want to fix later making a more generalized class.  Added similar styling selectors so using this originally absent mark is now consistent.  Now the default is to show all menu items/buttons.

More refactoring included:
- Better naming for the combination area of including the completed markers visibility control button, which did includes improving the semantic data, although even when other associated names have not been fixed yet.
- Also improved that styling to reduce the bottom margin to be more symmetrical with the top one.
- Hid action group links for other menus for now.
- Added missing config sample for new feature's already implemented property name.
- Some other light cleaning and organizing.
- Fixed problem accessing data from the URL that would be encoded and break code, such as the new feature having an array of ids, if spaces were included, and would not be automatically be URL-decoded or accepted as JSON when parsing later.
- Added cat content min height to show the back button better, for example, when I was fixing no menu items being shown at all, the arrow/chevron was only showing the upper half about.